Filterbank

The Filterbank module provides read/write access to SIGPROC Filterbank files (typically .fil files). A Filterbank file consists of a single header followed by a contiguous block of spectra. Each spectrum has nchans channels and nifs IFs (polarization products). Supported bit depths are 8 and 32.

Blio.Filterbank — Module

Module for interacting with Filterbank files.

See also: Header, Base.read!, Base.read, Base.write, Base.Array, chanfreq, chanfreqs, datasize, maskdc!

Headers

Filterbank.Header represents a SIGPROC Filterbank header. It is a subtype of AbstractDict{Symbol, Any} and behaves like an OrderedDict with the added ability to access header items as if they were struct fields.

julia> using Blio

julia> fbh = Filterbank.Header();

julia> fbh[:source_name] = "Demo"; fbh[:nchans] = 1024;

julia> fbh.source_name
"Demo"

julia> haskey(fbh, :fch1)
false

Construct an empty header with the default constructor, copy an existing header with the copy constructor, or build one from another Dict-like object or NamedTuple:

fbh = Filterbank.Header(other_fbh)
fbh = Filterbank.Header((source_name = "Demo", nchans = 1024))

You can also derive a Filterbank header from a GuppiRaw.Header:

fbh = Filterbank.Header(grh)
fbh = Filterbank.Header(grh; nbits = 8, source_name = "MySource")

Blio.Filterbank.Header — Type

Type used to hold a Filterbank header. Acts very much like an OrderedDict with the added ability to access header items as fields.

Blio.Filterbank.Header — Method

Header() -> Filterbank.Header
Header(other) -> Filterbank.Header

Construct an empty Filterbank.Header object or one initialized from other, which may be another Filterbank.Header, an AbstractDict{Symbol,Any}, or a NamedTuple.

Blio.Filterbank.Header — Method

Header(grh::GuppiRaw.Header; kwargs...) -> Filterbank.Header

Constructor to create Filterbank.Header from GuppiRaw.Header, translating corresponding fields as needed. Assumes Filterbank frequency channels and time samples have the same resolution as GuppiRaw counterparts. Fields may be overridden by passing them as kwargs. The following defaults are used:

telescope_id = -1

machine_id = -1

data_type = 1

barycentric = 0

pulsarcentric = 0

nbeams = 1

ibeam = 1

nbits = 32

nifs = 1

Reading and writing headers

Base.read! — Method

read!(io::IO, fbh::Filterbank.Header)::Filterbank.Header

Read and parse Filterbank header from io and populate fbh. Add header_size and data_size fields based on header size and file length.

If io was at the start of the file, it will be positioned at the start of data after this function returns. If io was positioned elsewhere, that position will remain unchanged after this function returns.

This function adds some additional unofficial "convenience" fields to the returned Filterbank.Header:

:header_size: number of bytes in header

:data_size: number of bytes of data

:sample_size: size of a single time sample (all channels, all IFs)

:nsamps: data size divided by sample size

The calculated :nsamps field should match any :nsamples field in the header. The :nsamples field is an optional header, but it is official so using that field for the derived "convenience" value is not desirable.

Base.read — Method

read(io::IO, Filterbank.Header)::Filterbank.Header

Create a Filterbank::Header object, then call read!() to populate it.

Base.write — Method

write(io::IO, fbh::Filterbank.Header; quiet=false)

Seeks to the beginning of io and writes a Filterbank header from the contents of fbh. If quiet is false (the default), then unknown/unsupported keywords will output warning messages.

Reading and writing individual header items

These lower-level helpers operate on a single keyword/value pair at a time. They are used internally by read! and write but are exported for callers who want fine-grained control.

Blio.Filterbank.read_header_item — Method

read_header_item(f::Function, io::IO)

Read a Filterbank header keyword and value from io. Call f(keyword, value, valpos), where keyword and value are from the header item and valpos is the file position from which the value was read. This can be used to update erroneous header values in place. Return (keyword, value).

Blio.Filterbank.read_header_item — Method

read_header_item(io::IO)

Call read_header_item with io and a no-op function for f. Return (keyword, value).

Blio.Filterbank.write_header_item — Function

write_header_item(io::IO, kw::Symbol, val=nothing; quiet=false)

Writes Filterbank header item kw and val to io. A value must be passed for all keywords other than :HEADER_START and :HEADER_END. If quiet is false (the default), then unknown/unsupported keywords will output warning messages.

Working with data

Memory mapping a Filterbank file

Blio.Filterbank.mmap — Method

mmap(fb::Union{IOStream,AbstractString})::Tuple{Filterbank.Header, Array}

Return a Filterbank.Header object and a memory mapped data Array for the data in fb.

Blio.Filterbank.mmap — Method

mmap(fb::Union{IOStream,AbstractString})::Tuple{Filterbank.Header, Array}
mmap(fb::IOStream, fbh::Filterbank.Header)::Array

Create a memory mapped Array for the data in fb. If Filterbank.Header fbh is passed, it will be used to determine Array type/size as well as the file position offset in fb and only the memory mapped data Array will be returned. Otherwise, a Filterbank.Header object will be created from the contents of fb and it will be returned along with the memory mapped data Array as a Tuple{Filterbank.Header, Array}.

The returned Array will be a ReinterpretArray if the file's data block is not type aligned.

Note

The underlying file handle will not be closed until the Array returned by Mmap.mmap is finalized. This will happen in due course via garbage collection, but it is also possible to call finalize to finalize the object earlier. Because Filterbank.mmap may return a ReinterpretArray, one must finalize the parent of the array returned by Filterbank.mmap if explicit finalization is desired. Finalizing the parent will work in all cases since a "regular" Array is its own parent.

Warning

After finaliztion, the data array is no longer valid. Using the data array after finalization will likely lead to segfaults!

Allocating an array

Core.Array — Method

Array(fbh::Filterbank.Header, nsamps::Int=0; <kwargs>)

Return an uninitialized Array sized for nsamps spectra of Filterbank data with dimensions derived from metadata in fbh, specifically the fbh.nchans, fbh.nifs, and fbh.nbits fields. The data type of the Array elements will be:

UInt8 when fbh.nbits == 8 and fbh.nifs == 1, or

Int8 when fbh.nbits == 8 and fbh.nifs != 1, or

Float32 when fbh.nbits == 32.

If nsamps is zero, the Array will be sized to hold all spectra from the file or as many spectra as will fit in maxmem bytes, whichever is less. The returned Array will hold at least one spectrum (assuming it gets successfully allocated), even if that would exceed maxmem. Files with exceptionally large numbers of channels may not be usable with this convention and the user will have to devise their own Array allocation scheme.

The Array will be dimensioned as (fbh.nchans, fbh.nifs, nsamps) unless nants > 1 (see below) or dropdims is true (in which case singleton dimensions will be eliminated).

Keyword Arguments

dropdims::Bool=false: drop singleton dimensions

maxmem::Int64=1<<32: limit nsamps to not more than this many bytes

nants::Int=1: If nants > 1, split the chan dimension into fbh.nchan÷nants and nants dimensions. It is an error if fbh.nchans is not a multiple of nants. The Array will be dimensioned as (fbh.nchans÷nants, nants, fbh.nifs, nsamps).

Dimensions and sizes

Blio.Filterbank.datasize — Function

datasize(fbh[, dim]; <kwargs>)

Return a tuple containing the dimensions of the data described by fbh. Optionally you can specify a dimension to just get the length of that dimension. This is not to be confused with fbh[:data_size] which is the byte size of the data portion of the Filterbank file.

fbh can be any Dict-like object that supports haskey and get, such as Filterbank.Header, AbstractDict, NamedTuple, DataFrameRow, etc.

Keyword Arguments

dropdims::Bool=false: drop singleton dimensions

nants::Int=1: If nants > 1, split the chan dimension into fbh.nchan÷nants and nants dimensions. It is an error if fbh.nchans is not a multiple of nants.

Masking DC channels

Blio.Filterbank.maskdc! — Function

maskdc!(fbdata::Array, ncoarse::Integer)::Nothing

Mask the center (aka "DC") fine channel of all the coarse channels that span the first dimension of fbdata by setting the DC fine channel to the mean of the two adjacent channels. The number of coarse channels in fbdata must be passed as ncoarse. The fbdata Array is modified in-place.

Channel frequencies

Blio.Filterbank.chanfreq — Method

chanfreq(fbh, chan::Real)::Float64

Returns the center frequency of the channel given by chan based on the fch1 and foff fields of fbh. The first channel in the file is considered to be channel 1 (i.e. chan is one-based).

fbh can be any Dict-like object that supports haskey and get, such as Filterbank.Header, AbstractDict, NamedTuple, DataFrameRow, etc.

Blio.Filterbank.chanfreqs — Method

chanfreqs(fbh, chans::AbstractRange=1:fbh.nchans)::AbstractRange

Returns the center frequencies of the channels given by chans based on the fch1, foff, and (in the default case) nchans fields of fbh. The first channel in the file is considered to be channel 1 (i.e. chans are one-based).

fbh can be any Dict-like object that supports haskey and get, such as Filterbank.Header, AbstractDict, NamedTuple, DataFrameRow, etc.

HDF5 conversion

The fil2h5 and h52fil functions are defined as empty stubs in the Filterbank module so that their docstrings are available even when HDF5.jl is not loaded. Their actual implementations are provided by the HDF5FilterbankExt extension module. See Extensions for details and the API documentation for fil2h5 and h52fil.

Full reference

Blio.Filterbank

Blio.Filterbank.Header

Blio.Filterbank.Header

Blio.Filterbank.Header

Core.Array

Base.read

Base.read!

Base.write

Blio.Filterbank.chanfreq

Blio.Filterbank.chanfreqs

Blio.Filterbank.datasize

Blio.Filterbank.maskdc!

Blio.Filterbank.mmap

Blio.Filterbank.mmap

Blio.Filterbank.read_header_item

Blio.Filterbank.read_header_item

Blio.Filterbank.write_header_item

This documentation was initially generated by opencode (powered by the openai/GLM-5.2 model) and subsequently reviewed and edited by the package author.