GuppiRaw

The GuppiRaw module provides read/write access to GUPPI Raw files (typically .raw files) produced by several radio telescope backends such as the GUPPI backend of the Green Bank Telescope (GBT) and the Breakthrough Listen backend of many telescopes (including the GBT). A GUPPI Raw file consists of a sequence of header + data block pairs. Each header is a series of 80-byte fixed-width records terminated by an END record (similar in spirit to a FITS header). The data block that follows is described by the header.

Blio.GuppiRaw — Module

Module for interacting with GuppiRaw files.

See also: Header, HeaderDataUnit, load, Base.read!, Base.read, Base.write, Base.Array, blocksize, blocktype, chanfreq, chanfreqs, getntime, resize_hdu, parse_header!

Headers

GuppiRaw.Header is the central type used to represent a GUPPI Raw header. It is a subtype of AbstractDict{Symbol, Any} and behaves much like an OrderedDict, but it also supports dot-syntax for keyword access:

julia> using Blio

julia> grh = GuppiRaw.Header();

julia> grh[:obsfreq] = 1400.0; grh[:obsnchan] = 64;

julia> grh.obsfreq
1400.0

julia> haskey(grh, :backend)
false

Construct an empty header with GuppiRaw.Header (the default constructor), or initialize one from an existing AbstractDict{Symbol, Any} or NamedTuple:

grh = GuppiRaw.Header((backend = "GUPPI", obsfreq = 1400.0, obsnchan = 64))

Blio.GuppiRaw.Header — Type

Type used to hold a GuppiRaw header

Reading and writing headers

Base.read! — Method

read!(io::IO, grh::GuppiRaw.Header; skip_padding::Bool=true)::Bool

Read a GUPPI header from io and populate grh.

Returns false if not enough data remain in io for a header, otherwise true is returned. This can be used to loop through a GuppiRaw file.

If skip_padding is true and grh[:directio] is non-zero, padding bytes will be skipped after the header such that the file position will be a multiple of 512 bytes. This will leave the file positioned at the start of the data block following the header. If reading a file of concatenated headers (i.e. without a data block after each header), use skip_padding=false.

Base.read — Method

read(io::IO, ::Type{GuppiRaw.Header};
     skip_padding::Bool=true)::GuppiRaw.Header

Create a GuppiRaw::Header object, call read!() to populate it, then return it.

Base.write — Method

write(io::IO, grh::GuppiRaw.Header; skip_padding::Bool=true)

Write grh as a GUPPI header to io. If skip_padding is true and grh[:directio] is non-zero, padding bytes will be written after the header such that the file position will be a multiple of 512 bytes.

Loading entire files

To read all of the header + data block pairs from a GUPPI Raw file (or set of files), use load. By default it returns a tuple of (headers, datablocks) where each data block is memory-mapped.

Blio.GuppiRaw.load — Function

load([predicate,] io::IO; headers=Header[], datablocks=Array{<:Complex{<:Integer}}[])
load([predicate,] filename; headers=Header[], datablocks=Array{<:Complex{<:Integer}}[])
load([predicate,] filenames; headers=Header[], datablocks=Array{<:Complex{<:Integer}}[])

The first form loads GUPPI RAW headers and datablocks from the current position of io to end of file. io must be positioned at the start of a GUPPI RAW header. The latter two forms read from the specified GUPPI RAW file or files.

The headers are read as GuppiRaw.Header objects and push!-ed onto headers. The data blocks are mmap-ed as Arrays and push!-ed onto datablocks. headers and datablocks default to empty Vectors, but any object onto which GuppiRaw.Headers or Arrays can be pushed may be given. If datablocks is given as false, then only the headers are read. This can be used to read the headers without reading or mmaping the data blocks.

If a predicate function is given it will be passed the first GuppiRaw.Header read from io or each file given. The predicate function can examine the Header to determine whether the file is considered valid (e.g. has expected sizing and/or frequency attributes). If the file is considered valid, the function should return true. If the file is not considered valid, the function may return false, which will cause load to return headers and datablocks unmodified for that file, or it may raise an exception if desired (which will also stop any further loading). Note that when multiple filenames are given, predicate is passed the first header of each file independently.

Returns the tuple (headers, datablocks), even when datablocks is false.

Data block layout

The shape and element type of a GUPPI Raw data block depend on the values of the header keywords BLOCSIZE, OBSNCHAN, NPOL, NANTS, and NBITS. The helper functions below query those values:

Blio.GuppiRaw.blocksize — Function

blocksize(grh[, dim])

Return a tuple containing the dimensions of the data block described by grh. Optionally you can specify a dimension to just get the length of that dimension.

If NANTS is 1 or unspecified, the returned tuple have three elements corresponding to (npol, ntime, obsnchan).

If NANTS is greater than 1, the returned tuple will have four elements corresponding to (npol, ntime, obsnchan÷nants, nants).

grh can be any Dict-like object that supports haskey and get, such as GuppiRaw.Header, AbstractDict, NamedTuple, DataFrameRow, etc.

Blio.GuppiRaw.blocktype — Function

blocktype(grh)

Return the type of an Array that corresponds to the data in the datablock described by grh. This will typically be one of the following types:

Array{Complex{Int8}, 3} when NBITS=8 and NANTS=1

Array{Complex{Int8}, 4} when NBITS=8 and NANTS>1

Array{Complex{Int16}, 3} when NBITS=16 and NANTS=1

Array{Complex{Int16}, 4} when NBITS=16 and NANTS>1

grh can be any Dict-like object that supports haskey and get, such as GuppiRaw.Header, AbstractDict, NamedTuple, DataFrameRow, etc.

Blio.GuppiRaw.getntime — Function

getntime(grh)::Int

Return the number of time samples per block.

grh can be any Dict-like object that supports haskey and get, such as GuppiRaw.Header, AbstractDict, NamedTuple, DataFrameRow, etc.

Blio.GuppiRaw.antnchan — Function

antnchan(grh)::Int

Returns the number of channels per antenna (i.e. obsnchan ÷ nants). Missing nants implies nants == 1.

grh can be any Dict-like object that supports haskey and get, such as GuppiRaw.Header, AbstractDict, NamedTuple, DataFrameRow, etc.

Allocating an array

To allocate a regular (non-memory-mapped) array sized appropriately for a data block described by grh, use Base.Array:

Core.Array — Method

Array(grh::GuppiRaw.Header, nchan::Int=0)::Array{Complex{Integer}}

Return an uninitialized 3 or 4 dimensional Array sized for nchan channels of RAW data as specified by metadata in header, specifically the BLOCSIZE, NANTS, OBSNCHAN, NPOL, and NBITS fields. nchan <= 0 implies all channels.

The data type of the Array elements will be Complex{Int8} when NBITS == 8 or Complex{Int16} when NBITS == 16.

If NANTS is 1 or unspecified, the Array will be dimensioned as [npol, ntime, nchan] to match the RAW data block layout.

If NANTS is greater than 1 and nchan is a multiple of OBSNCHAN/NANTS, then the returned array will be dimensioned as (npol, ntime, obsnchan/nants, nchan*nants/obschan), i.e. it will have an extra antenna dimension.

Memory mapping a single data block

Mmap.mmap — Method

mmap(io::IO, grh::Header) -> datablock

mmap a datablock starting from the current position of io. The type and size of the data block are as described by grh. The position of io is advanced by grh[:blocsize] (i.e. to just past the data block).

HeaderDataUnit

A HeaderDataUnit pairs a header with its data array:

Blio.GuppiRaw.HeaderDataUnit — Type

A HeaderDataUnit is a struct consisting of a Header and a RawArray

Blio.GuppiRaw.HeaderDataUnit — Method

HeaderDataUnit(grh::GuppiRaw.Header=GuppiRaw.Header())

Construct a HeaderDataUnit from a grh. If grh is empty, the data field will be Complex{Int8}[], otherwise it will be Array(grh). Note that grh is not copied, so when making multiple HeaderDataUnits be sure to use a different Header instance for each one.

Blio.GuppiRaw.resize_hdu — Function

resize_hdu(hdu::GuppiRaw.HeaderDataUnit)::GuppiRaw.HeaderDataUnit

Return a HeaderDataUnit whose data Array is appropriately typed and sized for the data block described by hdu.hdr. If hdu.data is already appropriately typed and sized then hdu is returned, otherwise a new HeaderDataUnit is returned with the same Header object as hdu.hdr but a new data Array that is appropriately typed and sized according to hdu.hdr.

Channel frequencies

Blio.GuppiRaw.chanfreq — Method

chanfreq(grh, chan::Real)::Float64

Returns the center frequency of the channel given by chan based on the obsfreq, chan_bw, obsnchan, and nants fields of grh. The first channel in the file is considered to be channel 1 (i.e. chan is one-based).

grh can be any Dict-like object that supports haskey and get, such as GuppiRaw.Header, AbstractDict, NamedTuple, DataFrameRow, etc.

Blio.GuppiRaw.chanfreqs — Method

chanfreqs(grh, chans::AbstractRange=1:grh[:obsnchan]/grh[:nants])::AbstractRange

Returns the center frequencies of the channels given by chans (or all channels if chans is not given) based on the obsfreq, chan_bw, obsnchan, and nants fields of grh. The first channel in the file is considered to be channel 1 (i.e. chans are one-based). Frequencies for channels beyond obsnchan/nants will be returned if requested, but they are not valid.

grh can be any Dict-like object that supports haskey and get, such as GuppiRaw.Header, AbstractDict, NamedTuple, DataFrameRow, etc.

Header parsing utilities

These functions are useful when working with raw header buffers, e.g. when inspecting headers that have not been read from a file.

Blio.GuppiRaw.parse_header! — Method

parse_header!(grh::Header, buf::Array{UInt8,2})

Find the END record in buf and parse the preceeding records into grh. The existing contents of grh are always removed, even if the END record is not found.

Blio.GuppiRaw.parse_header! — Method

parse_header!(grh::Header, buf::Array{UInt8,2}, endidx)

Parse the first endidx records in buf into grh (after removing any existing contents from grh). endidx should be the column index of the END record in buf. If endidx is nothing (e.g. if the END record does not exist), grh is still emptied.

Coordinate helpers

GUPPI Raw headers sometimes store the right ascension and declination of the source as strings (e.g. "HH:MM:SS.S") and sometimes as numeric values (in degrees). These helpers normalize both forms to a Float64 in degrees:

Blio.GuppiRaw.radeg — Function

radeg(ra)

Get ra as degrees. This field should be floating point degrees, but some files have it as a string in "H:M:S" form or a string in H.HHH... form. If the given ra is not a String, it will be returned unchanged. If ra is a String, it will be parsed as H:M:S and converted to degrees if it contains a colon (:), otherwise it will be parsed as being in floating point degrees format.

Blio.GuppiRaw.decdeg — Function

decdeg(dec)

Get dec as degrees. This field should be floating point degrees, but some files have it as a string in "D:M:S" form or a string in D.DDD... form. If the given dec is not a String, it will be returned unchanged. If dec is a String, it will be parsed as D:M:S and converted to degrees if it contains a colon (:), otherwise it will be parsed as being in floating point degrees format.

Full reference

Blio.GuppiRaw

Blio.GuppiRaw.Header

Blio.GuppiRaw.HeaderDataUnit

Blio.GuppiRaw.HeaderDataUnit

Core.Array

Base.read

Base.read!

Base.write

Blio.GuppiRaw.antnchan

Blio.GuppiRaw.blocksize

Blio.GuppiRaw.blocktype

Blio.GuppiRaw.chanfreq

Blio.GuppiRaw.chanfreqs

Blio.GuppiRaw.decdeg

Blio.GuppiRaw.getntime

Blio.GuppiRaw.load

Blio.GuppiRaw.parse_header!

Blio.GuppiRaw.parse_header!

Blio.GuppiRaw.radeg

Blio.GuppiRaw.resize_hdu

Mmap.mmap

This documentation was initially generated by opencode (powered by the openai/GLM-5.2 model) and subsequently reviewed and edited by the package author.